# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenFitLab API
  version: "1.0"
  description: |
    REST API for the OpenFitLab self-hosted fitness activity tracker.

    **Authentication:** All protected endpoints require a valid session cookie
    (`ofl.sid`). Obtain a session by completing the OAuth flow via
    `GET /api/auth/google` or `GET /api/auth/github`.

    **CSRF:** All state-changing requests (POST, PATCH, DELETE) must include the
    `x-csrf-token` header. Obtain the token from `GET /api/auth/me`.

    **Ownership:** All resource endpoints scope data to the authenticated user.
    A missing or unowned resource returns `404` (not `403`) to avoid leaking existence.

    **Error shape:** All error responses use `{ "error": "<message>" }`.
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Not found
    Stats:
      type: object
      description: |
        Key-value map of stat type to value. Common keys include Distance,
        Duration, Calories, HeartRateAvg, HeartRateMax, CadenceAvg, PowerAvg,
        SpeedAvg, ElevationGain. Values may be numbers, strings, or objects
        depending on stat type.
      additionalProperties: true
      example:
        Distance: 10000
        Duration: 3600
        HeartRateAvg: 145
    Activity:
      type: object
      required: [id, eventID, stats]
      properties:
        id:
          type: string
          format: uuid
        eventID:
          type: string
          format: uuid
        name:
          type: string
        startDate:
          type: integer
          description: Unix timestamp in milliseconds
        endDate:
          type: integer
          description: Unix timestamp in milliseconds
        type:
          type: string
          example: Running
        deviceName:
          type: string
          example: Garmin Forerunner 245
        startTimezone:
          type: string
          example: UTC
        endTimezone:
          type: string
          example: UTC
        stats:
          $ref: "#/components/schemas/Stats"
    Event:
      type: object
      required: [id, startDate, name, stats, folderId]
      properties:
        id:
          type: string
          format: uuid
        startDate:
          type: integer
          description: Unix timestamp in milliseconds
        endDate:
          type: integer
          description: Unix timestamp in milliseconds
        name:
          type: string
        description:
          type: string
        isMerge:
          type: boolean
        srcFileType:
          type: string
          example: tcx
        startTimezone:
          type: string
          example: UTC
        endTimezone:
          type: string
          example: UTC
        folderId:
          type: string
          format: uuid
          nullable: true
        stats:
          $ref: "#/components/schemas/Stats"
    EventDetail:
      type: object
      required: [event, activities]
      properties:
        event:
          $ref: "#/components/schemas/Event"
        activities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
    ActivityRow:
      type: object
      required: [event, activity]
      properties:
        event:
          $ref: "#/components/schemas/Event"
        activity:
          $ref: "#/components/schemas/Activity"
...
    StreamDataPoint:
      type: object
      required: [timeMs, value, sequenceIndex]
      properties:
        timeMs:
          type: integer
          description: Absolute timestamp in milliseconds
        value:
          description: Stream value (number, array, or object depending on stream type)
        sequenceIndex:
          type: integer
    Stream:
      type: object
      required: [id, type, activityId, dataPoints]
      properties:
        id:
          type: string
        type:
          type: string
          example: Heart Rate
        activityId:
          type: string
          format: uuid
        dataPoints:
          type: array
          items:
            $ref: "#/components/schemas/StreamDataPoint"
    UploadResultItem:
      type: object
      required: [filename, success]
      properties:
        filename:
          type: string
        success:
          type: boolean
        id:
          type: string
          format: uuid
          description: Present when success is true
        event:
          $ref: "#/components/schemas/Event"
          description: Present when success is true
        activities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
          description: Present when success is true. Activity ids are null in this response.
        error:
          type: string
          description: Present when success is false
...
  parameters:
    eventId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Event UUID
    activityId:
      name: activityId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Activity UUID
...
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found or not owned by the authenticated user
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Invalid or missing CSRF token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  securitySchemes:
    csrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
      description: CSRF token obtained from GET /api/auth/me. Required on all state-changing requests (POST, PATCH, DELETE).

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/events:
    get:
      tags: [events]
      summary: List events
      description: Returns enriched events (with activities and stats) in reverse-chronological order.
      parameters:
        - name: startDate
          in: query
          schema:
            type: integer
          description: Filter events with start_date >= startDate (ms)
        - name: endDate
          in: query
          schema:
            type: integer
          description: Filter events with start_date <= endDate (ms)
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: folderId
          in: query
          schema:
            type: string
          description: |
            Filter by folder. Use `unfiled` for events with no folder,
            `all` or omit for all events, or a UUID for a specific folder.
      responses:
        "200":
          description: Array of enriched events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [events]
      summary: Upload one or more activity files
      description: |
        Accepts 1–10 files as `multipart/form-data`. Each file is parsed,
        persisted, and returned in the results array. Unsupported or unparseable
        files produce a result with `success: false` rather than failing the
        whole request.
      security:
        - csrfToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                folderId:
                  type: string
                  format: uuid
                  description: Optional folder to assign all uploaded events to
      responses:
        "200":
          description: Upload results (one item per file)
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/UploadResultItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/events/activity-rows:
    get:
      tags: [events]
      summary: Paginated activity table
      description: Returns a paginated list of `{ event, activity }` pairs for the activity table view.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: startDate
          in: query
          schema:
            type: integer
        - name: endDate
          in: query
          schema:
            type: integer
        - name: activityTypes
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: devices
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: search
          in: query
          schema:
            type: string
          description: Partial match against event name, activity name, or activity type
        - name: folderId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated activity rows
          content:
            application/json:
              schema:
                type: object
                required: [rows, total]
                properties:
                  rows:
                    type: array
                    items:
                      $ref: "#/components/schemas/ActivityRow"
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/events/{id}:
    get:
      tags: [events]
      summary: Get event detail
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "200":
          description: Event with activities and stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags: [events]
      summary: Update event folder assignment
      security:
        - csrfToken: []
      parameters:
        - $ref: "#/components/parameters/eventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                folderId:
                  type: string
                  format: uuid
                  nullable: true
      responses:
        "204":
          description: Updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [events]
      summary: Delete event
      description: Deletes the event and any comparisons that reference it.
      security:
        - csrfToken: []
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/events/{id}/candidates:
    get:
      tags: [events]
      summary: Get comparison candidates
      description: |
        Returns events that overlap in time with the given event and are in the
        same folder (by default), suitable for building a new comparison.
      parameters:
        - $ref: "#/components/parameters/eventId"
        - name: sameFolderOnly
          in: query
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Overlapping candidate events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/events/{id}/activities/{activityId}:
    patch:
      tags: [events]
      summary: Update activity type or device name
      security:
        - csrfToken: []
      parameters:
        - $ref: "#/components/parameters/eventId"
        - $ref: "#/components/parameters/activityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: At least one of type or deviceName must be provided
              properties:
                type:
                  type: string
                  example: Cycling
                deviceName:
                  type: string
                  example: Wahoo ELEMNT Bolt
      responses:
        "200":
          description: Updated activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/events/{id}/activities/{activityId}/streams:
    get:
      tags: [events]
      summary: Get streams for an activity
      parameters:
        - $ref: "#/components/parameters/eventId"
        - $ref: "#/components/parameters/activityId"
        - name: types
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          description: Filter to specific stream types. Omit to return all streams.
      responses:
        "200":
          description: Streams with data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stream"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Comparisons ───────────────────────────────────────────────────────────
...